a program that print the first 50 fibonacci number, starting with 1 and 2

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,8 +8,8 @@
 
 int main(void)
 {
-	int sum;
-        int first, second;
+	long int sum;
+	long int first, second;
 	int count;
 
 	sum = 0;
